+++ src/net/lewisship/cli_tools/specs.clj
@@ -11,8 +11,7 @@
                                            ::groups
                                            ::transformer
                                            ::source-dirs
-                                           ::cache-dir
-                                           ::messy?]))
+                                           ::cache-dir]))
 
 (s/def ::non-blank-string (s/and string?
                                  #(not (str/blank? %))))
@@ -35,8 +34,6 @@
 (s/def ::transformer fn?)
 
 (s/def ::source-dirs (s/coll-of string?))
-
-(s/def ::messy? boolean?)
 
 ;; dispatch doesn't return
 (s/fdef cli-tools/dispatch :args (s/cat :options ::dispatch-options))